Refactored
- Usless files deleted
- File routes shortened
