+++ src/components/operator/OperatorInput.tsx
