+++ src/pages/operator/[id].tsx
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router'
-import operators from '../../components/operator/storage/OperatorStorage'
-import OperatorForm from '../../components/operator/forms/OperatorForm'
-import Footer from '../../components/footer/Footer'
+import operators from '../../components/operator/OperatorStorage'
+import OperatorForm from '../../components/operator/OperatorForm'
+import Footer from '../../components/Footer'
 
 const Payment = () => {
     const router = useRouter();
